requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
python-telegram-bot>=20.0
ngrok
python-dotenv
google-cloud-firestore